Explicitly enable WebMock in initializer
Fixes #43

When WebMock is not enabled in the spec_helper of the application being tested,
requests to the Stripe API are not intercepted by WebMock.

According to the WebMock documentation, in order to use WebMock outside a test
framework, `WebMock.enable!` needs to be called. Adding this to the WebMock
initializer makes WebMock actually work, and also fixes the build. I'm not sure
why the build started failing; there must have been a change to Stripe's API.