Update Stripe::Charge fixtures
This change updates the fixtures that create `Stripe::Charge` objects,
so that they confirm to the latest version of the API that we are using.
This gem will need to be updated for any of the other fixtures, as we
need to write tests for them.

Using this gem in the `#cw-stripe-service-specs` branch allows all of
the specs to pass, not that the charge object that is returned, matches
the expected public interface that the **real Stripe** returns.