Return error on native or zero amounts
When you pass a negative or 0 as an amount to Stripe it results in an
error being returned. On our app that uses FakeStripe, we had a bug
that was sending negative amounts to be charged and this was not caught
by the specs because FakeStripe was happily taking this request.

This introduces this error state to match the behavior of the Stripe
API. If a zero or negative amount is attemoted to be charged, an error
will be returned.